Major Changes and Updates
IMPORTANT: In this update, all database connection will be connected to
localhost database. To run the code properly, you must create a local
database and change the information in web.xml as required. Also, don't
forget to grant the user account the priviledge to create a datasource.

Other updates:
- Added logging to track bugs and misc.
- Change regular try-catch-finally to Java's new try-resources
implementation.
- Technician can now view all tickets and assign themselves to a ticket.
- NOTE: The difference between a technician and a supervisor now is only
a supervisor can assign other technicians to a ticket. In addition, only
supervisor can decline a ticket.
- Tickets now have a subject field.
- Department now will have a selection box rather than an input box.
- Added comments and code documentation in a few places that seems
confusing.
- Moved context file to the correct location.
- Change some front end UI elements.
- Edit some backend permission checking code to fix some major bugs.